refector : item test defaultOption 추가
