+++ src/test/java/com/prgrms/bdbks/domain/item/entity/ItemTest.java
@@ -1,6 +1,6 @@
 package com.prgrms.bdbks.domain.item.entity;
 
-import static com.prgrms.bdbks.domain.testutil.OrderObjectProvider.*;
+import static com.prgrms.bdbks.domain.testutil.ItemObjectProvider.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
@@ -20,6 +20,9 @@ class ItemTest {
 		String englishName = "Iced Caffe Americano";
 		int price = 4500;
 		String image = "https://hkbks.com/api/image.jpg";
+		DefaultOption defaultOption = createDefaultOption(1, 0, 0, 0,
+			BeverageOption.Milk.OAT, BeverageOption.Coffee.DECAFFEINATED,
+			BeverageOption.MilkAmount.MEDIUM);
 
 		//when & then
 		assertDoesNotThrow(() -> {
@@ -30,13 +33,14 @@ class ItemTest {
 				.price(price)
 				.image(image)
 				.description(description)
+				.defaultOption(defaultOption)
 				.build();
 		});
 	}
 
 	@DisplayName("생성 - Item() - 이름이 비어있으면 생성에 실패한다.")
 	@Test
-	void constructor_create_name_empty_fail() {
+	void constructor_nameEmpty_fail() {
 		//given
 		ItemCategory itemCategory = createReserveEspressoCategory();
 
@@ -45,6 +49,9 @@ class ItemTest {
 		String englishName = "Iced Caffe Americano";
 		int price = 4500;
 		String image = "https://hkbks.com/api/image.jpg";
+		DefaultOption defaultOption = createDefaultOption(1, 0, 0, 0,
+			BeverageOption.Milk.OAT, BeverageOption.Coffee.DECAFFEINATED,
+			BeverageOption.MilkAmount.MEDIUM);
 
 		//when & then
 		assertThrows(IllegalArgumentException.class, () -> {
@@ -55,13 +62,14 @@ class ItemTest {
 				.price(price)
 				.image(image)
 				.description(description)
+				.defaultOption(defaultOption)
 				.build();
 		});
 	}
 
 	@DisplayName("생성 - Item() - 한글명 글자 수가 30자 초과하면 생성에 실패한다.")
 	@Test
-	void constructor_create_with_name_length_over_30_fail() {
+	void constructor_createWithNameLengthOver30_fail() {
 		//given
 		ItemCategory itemCategory = createReserveEspressoCategory();
 
@@ -70,6 +78,9 @@ class ItemTest {
 		String englishName = "Iced Caffe Americano";
 		int price = 4500;
 		String image = "https://hkbks.com/api/image.jpg";
+		DefaultOption defaultOption = createDefaultOption(1, 0, 0, 0,
+			BeverageOption.Milk.OAT, BeverageOption.Coffee.DECAFFEINATED,
+			BeverageOption.MilkAmount.MEDIUM);
 
 		//when & then
 		assertThrows(IllegalArgumentException.class, () -> {
@@ -80,13 +91,14 @@ class ItemTest {
 				.price(price)
 				.image(image)
 				.description(description)
+				.defaultOption(defaultOption)
 				.build();
 		});
 	}
 
 	@DisplayName("생성 - Item() - 카테고리가 Null 이면 생성에 실패한다.")
 	@Test
-	void constructor_create_category_empty_fail() {
+	void constructor_createCategoryEmpty_fail() {
 		//given
 
 		String name = "아이스 아메리카노";
@@ -94,6 +106,9 @@ class ItemTest {
 		String englishName = "Iced Caffe Americano";
 		int price = 4500;
 		String image = "https://hkbks.com/api/image.jpg";
+		DefaultOption defaultOption = createDefaultOption(1, 0, 0, 0,
+			BeverageOption.Milk.OAT, BeverageOption.Coffee.DECAFFEINATED,
+			BeverageOption.MilkAmount.MEDIUM);
 
 		//when & then
 		assertThrows(NullPointerException.class, () -> {
@@ -103,13 +118,14 @@ class ItemTest {
 				.price(price)
 				.image(image)
 				.description(description)
+				.defaultOption(defaultOption)
 				.build();
 		});
 	}
 
 	@DisplayName("생성 - Item() - englishName이 Empty 이면 생성에 실패한다.")
 	@Test
-	void constructor_create_englishName_empty_fail() {
+	void constructor_createEnglishNameEmpty_fail() {
 		//given
 		ItemCategory itemCategory = createReserveEspressoCategory();
 
@@ -118,6 +134,9 @@ class ItemTest {
 		String englishName = null;
 		int price = 4500;
 		String image = "https://hkbks.com/api/image.jpg";
+		DefaultOption defaultOption = createDefaultOption(1, 0, 0, 0,
+			BeverageOption.Milk.OAT, BeverageOption.Coffee.DECAFFEINATED,
+			BeverageOption.MilkAmount.MEDIUM);
 
 		//when & then
 		assertThrows(IllegalArgumentException.class, () -> {
@@ -128,13 +147,14 @@ class ItemTest {
 				.price(price)
 				.image(image)
 				.description(description)
+				.defaultOption(defaultOption)
 				.build();
 		});
 	}
 
 	@DisplayName("생성 - Item() - price 가 0보다 작으면 생성에 실패한다.")
 	@Test
-	void constructor_create_price_less_then0_fail() {
+	void constructor_createPriceLessThen0_fail() {
 		//given
 		ItemCategory itemCategory = createReserveEspressoCategory();
 
@@ -143,6 +163,9 @@ class ItemTest {
 		String englishName = "Iced Caffe Americano";
 		int price = -1;
 		String image = "https://hkbks.com/api/image.jpg";
+		DefaultOption defaultOption = createDefaultOption(1, 0, 0, 0,
+			BeverageOption.Milk.OAT, BeverageOption.Coffee.DECAFFEINATED,
+			BeverageOption.MilkAmount.MEDIUM);
 
 		//when & then
 		assertThrows(IllegalArgumentException.class, () -> {
@@ -153,13 +176,14 @@ class ItemTest {
 				.price(price)
 				.image(image)
 				.description(description)
+				.defaultOption(defaultOption)
 				.build();
 		});
 	}
 
 	@DisplayName("생성 - Item() - description 이 empty 면 생성에 실패한다.")
 	@Test
-	void constructor_create_description_empty_fail() {
+	void constructor_createDescriptionEmpty_fail() {
 		//given
 		ItemCategory itemCategory = createReserveEspressoCategory();
 
@@ -168,6 +192,9 @@ class ItemTest {
 		String englishName = "Iced Caffe Americano";
 		int price = 4500;
 		String image = "https://hkbks.com/api/image.jpg";
+		DefaultOption defaultOption = createDefaultOption(1, 0, 0, 0,
+			BeverageOption.Milk.OAT, BeverageOption.Coffee.DECAFFEINATED,
+			BeverageOption.MilkAmount.MEDIUM);
 
 		//when & then
 		assertThrows(IllegalArgumentException.class, () -> {
@@ -178,25 +205,54 @@ class ItemTest {
 				.price(price)
 				.image(image)
 				.description(description)
+				.defaultOption(defaultOption)
 				.build();
 		});
 	}
 
 	@DisplayName("생성 - Item() - image 이 empty 거나 url이  아니라면 면 생성에 실패한다.")
 	@Test
-	void constructor_create_image_empty_fail() {
+	void constructor_createImageEmpty_fail() {
 		//given
 		ItemCategory itemCategory = createReserveEspressoCategory();
 
 		String name = "아이스 아메리카노";
 		String description = "진한 에스프레소에 시원한 정수물과 얼음을 더하여 스타벅스의 깔끔하고 강렬한 에스프레소를 가장 부드럽고 시원하게 즐길 수 있는 커피";
-
 		String englishName = "Iced Caffe Americano";
 		int price = 4500;
 		String image = "";
+		DefaultOption defaultOption = createDefaultOption(1, 0, 0, 0,
+			BeverageOption.Milk.OAT, BeverageOption.Coffee.DECAFFEINATED,
+			BeverageOption.MilkAmount.MEDIUM);
 
 		//when & then
 		assertThrows(IllegalArgumentException.class, () -> {
+			Item.builder()
+				.name(name)
+				.category(itemCategory)
+				.englishName(englishName)
+				.price(price)
+				.image(image)
+				.description(description)
+				.defaultOption(defaultOption)
+				.build();
+		});
+	}
+
+	@DisplayName("생성 - Item() - defaultOption이 Null 이면 생성에 실패한다.")
+	@Test
+	void constructor_createDefaultOptionNull_fail() {
+		//given
+		ItemCategory itemCategory = createReserveEspressoCategory();
+
+		String name = "아이스 아메리카노";
+		String description = "진한 에스프레소에 시원한 정수물과 얼음을 더하여 스타벅스의 깔끔하고 강렬한 에스프레소를 가장 부드럽고 시원하게 즐길 수 있는 커피";
+		String englishName = "Iced Caffe Americano";
+		int price = 4500;
+		String image = "https://hkbks.com/api/image.jpg";
+
+		//when & then
+		assertThrows(NullPointerException.class, () -> {
 			Item.builder()
 				.name(name)
 				.category(itemCategory)
